refactor(frontend): shared asset layer code
Move generic asset view layer code into `lib/data-map/layers/assets/`. Tooltips and data access are still handled by app config, as these depend on the network and hazards layer definitions.
