+++ frontend/src/lib/data-map/layers/assets/source.ts
